Moves CountryCodeString type
