+++ lib/fortnox/api/types/shim/country_code_string.rb
